Resolve some type issue.

+++ RagicViewer/classes/login/AZBasicAuthLoginViewController.m
@@ -169,7 +169,7 @@
     return cell;
 }
 
-- (float)tableView:(UITableView *)tableView heightForFooterInSection:(NSInteger)section {
+- (CGFloat)tableView:(UITableView *)tableView heightForFooterInSection:(NSInteger)section {
     // This will create a "invisible" footer
     return 40.0f;
 }
